pull out planningDataFor
